Fix single digit values are misinterpreted #4
The json file created by the "web-server" was always a string, which leads to some funky ordering (e.g., 9 > 50).

We can use the "value_type" from the database instead, if we are to support floats. But that should be in the series table, not in the sample in my opinion. So I think this can be tackled separatedly